Fixed some compiler errors 50597

+++ mz_compat.c
@@ -121,7 +121,7 @@ static int32_t mz_stream_ioapi_read(void *stream, void *buf, int32_t size) {
     } else
         return MZ_PARAM_ERROR;
 
-    return zread(opaque, ioapi->handle, buf, size);
+    return (int32_t)zread(opaque, ioapi->handle, buf, size);
 }
 
 static int32_t mz_stream_ioapi_write(void *stream, const void *buf, int32_t size) {
@@ -142,7 +142,7 @@ static int32_t mz_stream_ioapi_write(void *stream, const void *buf, int32_t size
     } else
         return MZ_PARAM_ERROR;
 
-    written = zwrite(opaque, ioapi->handle, buf, size);
+    written = (int32_t)zwrite(opaque, ioapi->handle, buf, size);
     return written;
 }
 
@@ -527,7 +527,7 @@ int zipOpenNewFileInZip3(zipFile file, const char *filename, const zip_fileinfo 
     unsigned long crc_for_crypting) {
     return zipOpenNewFileInZip3_64(file, filename, zipfi, extrafield_local, size_extrafield_local,
         extrafield_global, size_extrafield_global, comment, compression_method, level, raw, windowBits,
-        memLevel, strategy, password, crc_for_crypting, 0);
+        memLevel, strategy, password, (int32_t)crc_for_crypting, 0);
 }
 
 int zipOpenNewFileInZip3_64(zipFile file, const char *filename, const zip_fileinfo *zipfi,
@@ -592,7 +592,7 @@ int zipCloseFileInZipRaw64(zipFile file, uint64_t uncompressed_size, unsigned lo
     mz_compat *compat = (mz_compat *)file;
     if (!compat)
         return ZIP_PARAMERROR;
-    return mz_zip_entry_close_raw(compat->handle, (int64_t)uncompressed_size, crc32);
+    return mz_zip_entry_close_raw(compat->handle, (int64_t)uncompressed_size, (int32_t)crc32);
 }
 
 int zipCloseFileInZip(zipFile file) {
